move course_list remote, cache, data layers to course package
