+++ app/src/main/java/org/stepik/android/remote/course/source/CourseListRemoteDataSourceImpl.kt
@@ -1,14 +1,14 @@
-package org.stepik.android.remote.course_list
+package org.stepik.android.remote.course.source
 
 import io.reactivex.Single
 import org.stepic.droid.util.PagedList
-import org.stepik.android.data.course_list.source.CourseListRemoteDataSource
+import org.stepik.android.data.course.source.CourseListRemoteDataSource
 import org.stepik.android.domain.course_list.model.CourseListQuery
 import org.stepik.android.model.Course
 import org.stepik.android.remote.base.mapper.toPagedList
 import org.stepik.android.remote.course.model.CourseResponse
-import org.stepik.android.remote.course_list.mapper.CourseListQueryMapper
-import org.stepik.android.remote.course_list.service.CourseListService
+import org.stepik.android.remote.course.mapper.CourseListQueryMapper
+import org.stepik.android.remote.course.service.CourseListService
 import javax.inject.Inject
 
 class CourseListRemoteDataSourceImpl
