+++ app/src/main/java/org/stepik/android/view/injection/course/CourseComponent.kt
@@ -9,7 +9,6 @@ import org.stepik.android.view.course_info.ui.fragment.CourseInfoFragment
 import org.stepik.android.view.course_reviews.ui.fragment.CourseReviewsFragment
 import org.stepik.android.view.injection.calendar.CalendarDataModule
 import org.stepik.android.view.injection.course_calendar.CourseCalendarDataModule
-import org.stepik.android.view.injection.course_list.CourseListDataModule
 import org.stepik.android.view.injection.course_payments.CoursePaymentsDataModule
 import org.stepik.android.view.injection.course_reviews.CourseReviewsDataModule
 import org.stepik.android.view.injection.last_step.LastStepDataModule
@@ -30,7 +29,6 @@ import org.stepik.android.view.injection.user.UserDataModule
     UserDataModule::class,
     StepDataModule::class,
 
-    CourseListDataModule::class,
     CoursePaymentsDataModule::class,
     CourseReviewsDataModule::class,
     PersonalDeadlinesDataModule::class,
